test: 로컬 환경에서 테스트 하기 위한 시스템 구축
act 와 docker 를 활용해서 github action 을 로컬에서 동작함. git 관련 동작 (push, commit, issue 발생) 등은 제외하고 스크래핑과 자바스크립트로 이슈 문자열 만드는 과정 테스트
