+++ .github/actions/create-blog-post-issue-test/action.yaml
@@ -1,0 +1,9 @@
+name: 'Create Blog Post Issue Test'
+description: 'Creates an issue in GitHub for new blog posts.'
+runs:
+  using: 'node20'
+  main: 'create-issue_test.js'
+inputs:
+  POST_DATA:
+    description: 'Data of the new blog posts to include in the issue.'
+    required: true
