+++ .github/workflows/main_test.yaml
@@ -1,12 +1,11 @@
-name: 새로운 블로그 게시글 확인
+---
+name: 새로운 블로그 게시글 확인 테스트
 
 on:
-  schedule:
-    - cron: '0 23 * * *' # 한국(UTC+9) 기준 오전 8시
   workflow_dispatch:
 
 jobs:
-  check_blog_post: # action 메인 job
+  check_blog_post_test: # action 메인 job
     runs-on: ubuntu-latest
     steps:
     - name: Checkout code
@@ -22,9 +21,6 @@ jobs:
       with:
         python-version: '3.10'
     
-    # - name: Setup Java JDK
-    #   uses: actions/setup-java@v4.2.1
-                
     - name: Install python dependencies
       run: |
         python -m pip install --upgrade pip virtualenv
@@ -34,7 +30,7 @@ jobs:
 
     - name: Install Node.js dependencies
       run: |
-        cd ./.github/actions/create-blog-post-issue
+        cd ./.github/actions/create-blog-post-issue-test
         npm install
 
     - name: Check for new blog posts and update text files
@@ -47,19 +43,10 @@ jobs:
         else
           echo "POST_DATA=$output" >> $GITHUB_ENV
         fi
-    
-    - name: Commit and push changes if new posts are found
-      if: env.POST_DATA != ''
-      run: |
-        git config --global user.name 'dgh06175'
-        git config --global user.email 'dgh06175@gmail.com'
-        git add pastData/*.txt
-        git commit -m "action: 이슈 발생한 게시글 주소 텍스트 파일에 추가"
-        git push
 
     - name: Create GitHub Issue for new posts
       if: env.POST_DATA != ''
-      uses: ./.github/actions/create-blog-post-issue
+      uses: ./.github/actions/create-blog-post-issue-test
       with:
-        GITHUB_TOKEN: ${{ secrets.TECHBLOG_ALARM_ISSUE_ACTION_TOKEN }}
         POST_DATA: ${{ env.POST_DATA }}
+...
